🔧 [build] Enable Windows/MSVC builds
Add Windows platform support with MSVC compiler profiles, build system
adjustments, and updated documentation.

Build system:
- Add Conan profiles for Windows (x86_64, Ninja) and MSVC (v195, C++23)
- Add MSVC warning suppressions: C5030 (vendor attributes like
  [[gnu::pure]]) and C4324 (intentional cacheline padding)
- Handle MSVC-specific sanitizer flags (/fsanitize=address) and skip
  unsupported sanitizers (ubsan, tsan, lsan) with warnings
- Replace symlink with Windows junction for build directory shortcut
- Fix implicit double-to-uint64 conversion warning in clock::ticks

Source fixes:
- Add missing <compare> include for MSVC in snapshot.hpp (GCC/Clang
  include it transitively, MSVC does not)
- Reformat trailing return types in clock.hpp

IDE:
- Make clangd path platform-agnostic and remove Linux-only --malloc-trim

Documentation:
- Add Windows/MSVC build instructions to README.md and CONTRIBUTING.md
- Document Visual Studio 2026 Build Tools prerequisite
- Add Windows examples for address sanitizer and sanity checks
- Note coverage as Linux-only (lcov/gcov)
- Update test framework references from GUnit to GoogleTest

+++ include/jage/time/internal/clock.hpp
@@ -25,10 +25,11 @@ class clock {
   std::uint64_t elapsed_ticks_{};
   double time_scale_{1.0};
 
-  [[nodiscard]] auto
-  ticks(const duration_ current_time) const -> std::uint64_t {
+  [[nodiscard]] auto ticks(const duration_ current_time) const
+      -> std::uint64_t {
     const auto accumulated_time = current_time - elapsed_time_;
-    return std::floor(accumulated_time.count() / tick_duration_.count()) +
+    return static_cast<std::uint64_t>(
+               std::floor(accumulated_time.count() / tick_duration_.count())) +
            elapsed_ticks_;
   }
 
@@ -51,8 +52,8 @@ public:
     return duration_type{ticks() * tick_duration_.count()};
   }
 
-  [[nodiscard]] constexpr auto
-  tick_duration() const noexcept -> const duration_type & {
+  [[nodiscard]] constexpr auto tick_duration() const noexcept
+      -> const duration_type & {
     return tick_duration_;
   }
